move the index worker to its own file
